Pass to eliminate init loops (#12)
An example of a simple transformation pass that doesn't use patterns, but a more "classic" walk across the IR.

The purpose of the pass is to eliminate for-loops that initialize an empty tensor with the loop induction value in every iteration and can be replaced by a tensor range initialization under a number of conditions.

The pass also defines a debugging type and a pass statistic counting the number of loops eliminated.

---------

Signed-off-by: Lukas Sommer <lukas.sommer@codeplay.com>